+++ vantage-diorama/src/ops/mod.rs
@@ -1,7 +1,9 @@
 pub mod change_event;
 pub mod change_flash;
+pub mod flash_rejection;
 pub mod query_descriptor;
 
 pub use change_event::ChangeEvent;
 pub use change_flash::{ChangeFlash, FlashKind};
+pub use flash_rejection::FlashRejection;
 pub use query_descriptor::QueryDescriptor;